Projection with data through 3/03/2021
I replaced the data going back to the series start with more accurate values from Bloomberg's archive files on Github. This resulted in dropping a few days at the start of the series in December 2020, and adjusting the trend calculations accordingly. Projections were rerun for all days from 1/22/2021 to 3/02/2021 for purposes of historical data retention.